Switch update-issue to node script
